requirements.txt
deta<=1.2.0
google_api_python_client<=2.97.0
httplib2<=0.22.0
numpy<=1.24.3
oauth2client<=4.1.3
paddleocr<=2.7.0.2
pandas<=1.5.3
Pillow<=9.3.0
Pillow<=10.0.0
python-dotenv<=1.0.0
streamlit<=1.25.0
streamlit_elements<=0.1.0
streamlit_img_label<=0.1.1
streamlit_javascript<=0.1.5
streamlit_sparrow_labeling<=0.1.1
